refactor: 金額欄位統一為 BigDecimal + DECIMAL(10,2)
金額不該用 Double/DOUBLE PRECISION（浮點精度問題），
統一使用 BigDecimal (Java) + DECIMAL(10,2) (DB)。

變更：
- Plan entity: priceMonthly, priceYearly, priceUsdt → BigDecimal
- PaymentHistory entity: amount → BigDecimal
- DTOs: PlanResponse, CryptoCheckoutResponse 對應欄位 → BigDecimal
- SubscriptionService: 移除 Double↔BigDecimal 轉換，直接傳遞 BigDecimal
- V16 migration: ALTER plans/payment_history 金額欄位為 DECIMAL(10,2)
- 測試：斷言改用 isEqualByComparingTo() 比較 BigDecimal

+++ src/main/java/com/trader/subscription/dto/CryptoCheckoutResponse.java
@@ -3,6 +3,8 @@ package com.trader.subscription.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * USDT 付款資訊回應
  *
@@ -13,7 +15,7 @@ import lombok.Data;
 public class CryptoCheckoutResponse {
     private String planId;
     private String planName;
-    private Double amountUsdt;
+    private BigDecimal amountUsdt;
     private String walletAddress;
     private String network;
 }
